+++ TeamCode/src/test/java/com/acmerobotics/relicrecovery/path/PathFollowerTest.java
@@ -15,18 +15,18 @@ import static org.mockito.Mockito.mock;
  * @author Ryan
  */
 
-public class LinearPathExecutorTest {
+public class PathFollowerTest {
 
     @Test
     public void testBasicPathExecution() {
         MecanumDrive drive = mock(MecanumDrive.class);
 
-        LinearPath path = new LinearPath(Arrays.asList(
+        Path path = new Path(Arrays.asList(
                 new Pose2d(0, 0),
                 new Pose2d(2, 0),
                 new Pose2d(2, 2)
         ));
-        LinearPathExecutor executor = new LinearPathExecutor(drive);
+        PathFollower executor = new PathFollower(drive);
         executor.execute(path, 0.5, 0);
 
         InOrder inOrder = inOrder(drive);
